+++ src/components/Restaurant/component.jsx
@@ -1,6 +1,6 @@
 import classNames from 'classnames';
 import { MenuContainer } from '../Menu/container';
-import { Reviews } from '../Reviews/component';
+import { ReviewsContainer } from '../Reviews/container';
 
 export const Restaurant = ({ restaurant }) => {
 	return (
@@ -8,7 +8,7 @@ export const Restaurant = ({ restaurant }) => {
       ? <div className={classNames('margin-l-2')}>
         <h2>{restaurant.name}</h2>
         <MenuContainer restaurantId={restaurant.id} />
-        <Reviews restaurantId={restaurant.id}/>
+        <ReviewsContainer restaurantId={restaurant.id}/>
       </div>
       : null
 	)
