--- src/redux/entities/restaurants/thunks/get-restaurants.js
@@ -1,10 +1,0 @@
-import { createAsyncThunk } from "@reduxjs/toolkit";
-
-export const getRestaurants = createAsyncThunk(
-  'restaurants/getRestaurants',
-  async () => {
-    const response = fetch('http://localhost:3001/api/restaurants/')
-
-    return (await response).json();
-  }
-);
