+++ src/redux/index.js
@@ -1,7 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import dish from './entities/dish';
 import restaurant from './entities/restaurant';
-import restaurants from './entities/restaurants';
 import review from './entities/review';
 import user from './entities/user';
 import { loggerMiddleware } from './middlewares/logger';
@@ -10,7 +9,6 @@ const store = configureStore({
   reducer: {
     dish,
     restaurant,
-    restaurants,
     review,
     user,
   },
